+++ R/qc_peptide_type.R
@@ -1,20 +1,29 @@
 #' Check peptide type percentage share
 #'
-#' Calculates the percentage share of each peptide types (fully-tryptic, semi-tryptic, non-tryptic) for each sample.
+#' Calculates the percentage share of each peptide types (fully-tryptic, semi-tryptic,
+#' non-tryptic) for each sample.
 #'
-#' @param data A data frame containing at least the input columns.
+#' @param data a data frame containing at least the input columns.
 #' @param sample the name of the column containing the sample names.
 #' @param peptide the name of the column containing the peptide sequence.
-#' @param pep_type the name of the column containing the peptide type. Can be obtained using the \code{find_peptide} and \code{peptide_type} function together.
-#' @param intensity the name of the column containing the corresponding raw or normalised intensity values (not log2) for each peptide or precursor. Required when "intensity" is chosen as the method.
-#' @param remove_na_intensities Logical specifying if sample/peptide combinations with intensities that are NA (not quantified IDs) should
-#' be dropped from the data frame for analysis of peptide type distributions. Default is TRUE since we are usually
-#' interested in the peptide type distribution of quantifiable IDs. This is only relevant for method = "count".
-#' @param method character vector indicating the method used for evaluation. \code{method = "intensity"} calculates the peptide type percentage by intensity, whereas \code{method = "count"} calculates the percentage by peptide ID count. Default is \code{method = count}.
+#' @param pep_type the name of the column containing the peptide type. Can be obtained using the
+#' \code{find_peptide} and \code{assign_peptide_type} function together.
+#' @param intensity the name of the column containing the corresponding raw or normalised
+#' intensity values (not log2) for each peptide or precursor. Required when "intensity" is
+#' chosen as the method.
+#' @param remove_na_intensities Logical specifying if sample/peptide combinations with intensities
+#' that are NA (not quantified IDs) should be dropped from the data frame for analysis of peptide
+#' type distributions. Default is TRUE since we are usually interested in the peptide type
+#' distribution of quantifiable IDs. This is only relevant for method = "count".
+#' @param method character vector indicating the method used for evaluation.
+#' \code{method = "intensity"} calculates the peptide type percentage by intensity, whereas
+#' \code{method = "count"} calculates the percentage by peptide ID count. Default is \code{method = count}.
 #' @param plot a logical indicating whether the result should be plotted.
 #' @param interactive a logical indicating whether the plot should be interactive.
 #'
-#' @return A data frame that contains the calculated percentage shares of each peptide type per sample. The \code{count} column contains the number of peptides with a specific type. The \code{peptide_type_percent} column contains the percentage share of a specific peptide type.
+#' @return A data frame that contains the calculated percentage shares of each peptide type per
+#' sample. The \code{count} column contains the number of peptides with a specific type. The
+#' \code{peptide_type_percent} column contains the percentage share of a specific peptide type.
 #' @import dplyr
 #' @import tidyr
 #' @import ggplot2