+++ tests/testthat/test-auxiliary_functions.R
@@ -22,23 +22,56 @@ if (Sys.getenv("TEST_PROTTI") == "true") {
       )
     )
 
-    assigned_types <- data %>%
-      find_peptide(
-        protein_sequence = protein_sequence,
-        peptide_sequence = peptide
-      ) %>%
-      peptide_type(aa_before = aa_before, last_aa = last_aa)
-
-    test_that("find_peptide and peptide_type work", {
+    test_that("deprecated peptide_type works", {
+      rlang::with_options(lifecycle_verbosity = "warning", {
+        expect_warning(assigned_types <- data %>%
+          find_peptide(
+            protein_sequence = protein_sequence,
+            peptide_sequence = peptide
+          ) %>%
+          peptide_type(aa_before = aa_before, last_aa = last_aa))
+      })
       expect_is(assigned_types, "data.frame")
       expect_equal(nrow(assigned_types), 3)
       expect_equal(ncol(assigned_types), 9)
       expect_equal(assigned_types$pep_type, c("fully-tryptic", "semi-tryptic", "non-tryptic"))
     })
 
-    coverage <- sequence_coverage(data = assigned_types, protein_sequence = protein_sequence, peptides = peptide)
+    assigned_types <- data %>%
+      find_peptide(
+        protein_sequence = protein_sequence,
+        peptide_sequence = peptide
+      ) %>%
+      assign_peptide_type(aa_before = aa_before, last_aa = last_aa)
 
-    test_that("sequence_coverage works", {
+    test_that("find_peptide and assign_peptide_type work", {
+      expect_is(assigned_types, "data.frame")
+      expect_equal(nrow(assigned_types), 3)
+      expect_equal(ncol(assigned_types), 9)
+      expect_equal(assigned_types$pep_type, c("fully-tryptic", "semi-tryptic", "non-tryptic"))
+    })
+
+    test_that("deprecated sequence_coverage works", {
+      rlang::with_options(lifecycle_verbosity = "warning", {
+        expect_warning(coverage <- sequence_coverage(
+          data = assigned_types,
+          protein_sequence = protein_sequence,
+          peptides = peptide
+        ))
+      })
+      expect_is(coverage, "data.frame")
+      expect_equal(nrow(coverage), 3)
+      expect_equal(ncol(coverage), 10)
+      expect_equal(unique(round(coverage$coverage, digits = 1)), 7.7)
+    })
+
+    coverage <- calculate_sequence_coverage(
+      data = assigned_types,
+      protein_sequence = protein_sequence,
+      peptides = peptide
+    )
+
+    test_that("calculate_sequence_coverage works", {
       expect_is(coverage, "data.frame")
       expect_equal(nrow(coverage), 3)
       expect_equal(ncol(coverage), 10)