+++ tests/testthat/test-workflow.R
@@ -20,13 +20,53 @@ data_drc <- create_synthetic_data(
   additional_metadata = FALSE
 )
 
+if (Sys.getenv("TEST_PROTTI") == "true") {
+  test_that("deprecated median_normalization works", {
+    rlang::with_options(lifecycle_verbosity = "warning", {
+      expect_warning(normalised_data <- data %>%
+        median_normalisation(sample = sample, intensity_log2 = peptide_intensity_missing))
+    })
+    rlang::with_options(lifecycle_verbosity = "warning", {
+      expect_warning(normalised_data_drc <- data_drc %>%
+        median_normalisation(sample = sample, intensity_log2 = peptide_intensity_missing))
+    })
+
+    non_normalised <- normalised_data %>%
+      dplyr::group_by(sample) %>%
+      dplyr::summarise(median = median(peptide_intensity_missing, na.rm = TRUE), .groups = "drop")
+    normalised <- normalised_data %>%
+      dplyr::group_by(sample) %>%
+      dplyr::summarise(median = median(normalised_intensity_log2, na.rm = TRUE), .groups = "drop")
+
+    all_equal_non_normalised <- range(non_normalised$median) / mean(non_normalised$median)
+    expect_false(isTRUE(all.equal(all_equal_non_normalised[1], all_equal_non_normalised[2]))) # test that medians are unequal before normalizing
+
+    all_equal_normalised <- range(normalised$median) / mean(normalised$median)
+    expect_equal(all_equal_normalised[1], all_equal_normalised[2]) # test that medians are equal after normalizing
+
+    ## test drc data
+    non_normalised_drc <- normalised_data_drc %>%
+      dplyr::group_by(sample) %>%
+      dplyr::summarise(median = median(peptide_intensity_missing, na.rm = TRUE), .groups = "drop")
+    normalised_drc <- normalised_data_drc %>%
+      dplyr::group_by(sample) %>%
+      dplyr::summarise(median = median(normalised_intensity_log2, na.rm = TRUE), .groups = "drop")
+
+    all_equal_non_normalised_drc <- range(non_normalised_drc$median) / mean(non_normalised_drc$median)
+    expect_false(isTRUE(all.equal(all_equal_non_normalised_drc[1], all_equal_non_normalised_drc[2]))) # test that medians are unequal before normalizing
+
+    all_equal_normalised_drc <- range(normalised_drc$median) / mean(normalised_drc$median)
+    expect_equal(all_equal_normalised_drc[1], all_equal_normalised_drc[2]) # test that medians are equal after normalizing
+  })
+}
+
 normalised_data <- data %>%
-  median_normalisation(sample = sample, intensity_log2 = peptide_intensity_missing)
+  normalise(sample = sample, intensity_log2 = peptide_intensity_missing, method = "median")
 
 normalised_data_drc <- data_drc %>%
-  median_normalisation(sample = sample, intensity_log2 = peptide_intensity_missing)
+  normalise(sample = sample, intensity_log2 = peptide_intensity_missing, method = "median")
 
-test_that("median_normalization works", {
+test_that("normalise works", {
   non_normalised <- normalised_data %>%
     dplyr::group_by(sample) %>%
     dplyr::summarise(median = median(peptide_intensity_missing, na.rm = TRUE), .groups = "drop")
@@ -103,9 +143,43 @@ test_that("calculate_protein_abundance works", {
   expect_equal(ncol(protein_abundance_all), 4)
 })
 
-test_that("plot_peptide_profiles works", {
+if (Sys.getenv("TEST_PROTTI") == "true") {
+  test_that("deprecated plot_peptide_profiles works", {
+    # not testing split_all = TRUE. Also not testing protein_abundance_plot = FALSE.
+    expect_warning(p <- plot_peptide_profiles(
+      data = protein_abundance_all,
+      sample = sample,
+      peptide = peptide,
+      intensity_log2 = normalised_intensity_log2,
+      grouping = protein,
+      targets = c("protein_1", "protein_2"),
+      protein_abundance_plot = TRUE
+    ))
+    expect_is(p, "list")
+    expect_is(p[[1]], "ggplot")
+    expect_error(print(p[[1]]), NA)
+
+    rlang::with_options(lifecycle_verbosity = "warning", {
+      expect_warning(p_peptides <- plot_peptide_profiles(
+        data = missing_data,
+        sample = sample,
+        peptide = peptide,
+        intensity_log2 = normalised_intensity_log2,
+        grouping = protein,
+        targets = c("protein_12"),
+        protein_abundance_plot = FALSE
+      ))
+    })
+
+    expect_is(p_peptides, "list")
+    expect_is(p_peptides[[1]], "ggplot")
+    expect_error(print(p_peptides[[1]]), NA)
+  })
+}
+
+test_that("peptide_profile_plot works", {
   # not testing split_all = TRUE. Also not testing protein_abundance_plot = FALSE.
-  p <- plot_peptide_profiles(
+  p <- peptide_profile_plot(
     data = protein_abundance_all,
     sample = sample,
     peptide = peptide,
@@ -119,7 +193,7 @@ test_that("plot_peptide_profiles works", {
   expect_error(print(p[[1]]), NA)
 
   if (Sys.getenv("TEST_PROTTI") == "true") {
-    p_peptides <- plot_peptide_profiles(
+    p_peptides <- peptide_profile_plot(
       data = missing_data,
       sample = sample,
       peptide = peptide,
@@ -170,41 +244,101 @@ test_that("calculate_diff_abundance works", {
 
 if (Sys.getenv("TEST_PROTTI") == "true") {
   test_that("deprecated diff_abundance works", {
-    diff_deprecated <- diff_abundance(data = missing_data, sample = sample, condition = condition, grouping = peptide, intensity_log2 = normalised_intensity_log2, missingness = missingness, comparison = comparison, method = "t-test", retain_columns = c(protein))
+    rlang::with_options(lifecycle_verbosity = "warning", {
+      expect_warning(diff_deprecated <- diff_abundance(
+        data = missing_data,
+        sample = sample,
+        condition = condition,
+        grouping = peptide,
+        intensity_log2 = normalised_intensity_log2,
+        missingness = missingness,
+        comparison = comparison,
+        method = "t-test",
+        retain_columns = c(protein)
+      ))
+    })
     expect_is(diff_deprecated, "data.frame")
     expect_equal(nrow(diff_deprecated), 601)
     expect_equal(ncol(diff_deprecated), 9)
     expect_equal(round(min(diff_deprecated$adj_pval, na.rm = TRUE), digits = 9), 0.00758761)
 
-    if (Sys.getenv("TEST_PROTTI") == "true") {
-      data_mean_sd <- missing_data %>%
-        tidyr::drop_na() %>%
-        dplyr::group_by(condition, peptide, protein) %>%
-        dplyr::summarise(mean = mean(normalised_intensity_log2, na.rm = TRUE), sd = sd(normalised_intensity_log2, na.rm = TRUE), n = dplyr::n(), .groups = "drop")
+    data_mean_sd <- missing_data %>%
+      tidyr::drop_na() %>%
+      dplyr::group_by(condition, peptide, protein) %>%
+      dplyr::summarise(mean = mean(normalised_intensity_log2, na.rm = TRUE), sd = sd(normalised_intensity_log2, na.rm = TRUE), n = dplyr::n(), .groups = "drop")
 
-      diff_mean_sd_deprecated <- diff_abundance(data = data_mean_sd, condition = condition, grouping = peptide, mean = mean, sd = sd, n_samples = n, ref_condition = "condition_1", method = "t-test_mean_sd", retain_columns = c(protein))
-      diff_moderated_deprecated <- diff_abundance(data = missing_data, sample = sample, condition = condition, grouping = peptide, intensity_log2 = normalised_intensity_log2, missingness = missingness, comparison = comparison, method = "moderated_t-test", retain_columns = c(protein))
-      diff_proDA_deprecated <- diff_abundance(data = missing_data, sample = sample, condition = condition, grouping = peptide, intensity_log2 = normalised_intensity_log2, missingness = missingness, comparison = comparison, method = "proDA", retain_columns = c(protein))
+    rlang::with_options(lifecycle_verbosity = "warning", {
+      expect_warning(diff_mean_sd_deprecated <- diff_abundance(
+        data = data_mean_sd,
+        condition = condition,
+        grouping = peptide,
+        mean = mean,
+        sd = sd,
+        n_samples = n,
+        ref_condition = "condition_1",
+        method = "t-test_mean_sd",
+        retain_columns = c(protein)
+      ))
+    })
+    rlang::with_options(lifecycle_verbosity = "warning", {
+      expect_warning(diff_moderated_deprecated <- diff_abundance(
+        data = missing_data,
+        sample = sample,
+        condition = condition,
+        grouping = peptide,
+        intensity_log2 = normalised_intensity_log2,
+        missingness = missingness,
+        comparison = comparison,
+        method = "moderated_t-test",
+        retain_columns = c(protein)
+      ))
+    })
+    rlang::with_options(lifecycle_verbosity = "warning", {
+      expect_warning(diff_proDA_deprecated <- diff_abundance(
+        data = missing_data,
+        sample = sample,
+        condition = condition,
+        grouping = peptide,
+        intensity_log2 = normalised_intensity_log2,
+        missingness = missingness,
+        comparison = comparison,
+        method = "proDA",
+        retain_columns = c(protein)
+      ))
+    })
 
-      expect_is(diff_mean_sd_deprecated, "data.frame")
-      expect_is(diff_moderated_deprecated, "data.frame")
-      expect_is(diff_proDA_deprecated, "data.frame")
-      expect_equal(nrow(diff_mean_sd_deprecated), 599)
-      expect_equal(nrow(diff_moderated_deprecated), 601)
-      expect_equal(nrow(diff_proDA_deprecated), 601)
-      expect_equal(ncol(diff_mean_sd_deprecated), 14)
-      expect_equal(ncol(diff_moderated_deprecated), 13)
-      expect_equal(ncol(diff_proDA_deprecated), 12)
-      expect_equal(round(min(diff_mean_sd_deprecated$adj_pval, na.rm = TRUE), digits = 9), 0.00758761)
-      expect_equal(round(min(diff_moderated_deprecated$adj_pval, na.rm = TRUE), digits = 9), 5.1129e-05)
-      expect_equal(round(min(diff_proDA_deprecated$adj_pval, na.rm = TRUE), digits = 5), 0.00125)
-    }
+    expect_is(diff_mean_sd_deprecated, "data.frame")
+    expect_is(diff_moderated_deprecated, "data.frame")
+    expect_is(diff_proDA_deprecated, "data.frame")
+    expect_equal(nrow(diff_mean_sd_deprecated), 599)
+    expect_equal(nrow(diff_moderated_deprecated), 601)
+    expect_equal(nrow(diff_proDA_deprecated), 601)
+    expect_equal(ncol(diff_mean_sd_deprecated), 14)
+    expect_equal(ncol(diff_moderated_deprecated), 13)
+    expect_equal(ncol(diff_proDA_deprecated), 12)
+    expect_equal(round(min(diff_mean_sd_deprecated$adj_pval, na.rm = TRUE), digits = 9), 0.00758761)
+    expect_equal(round(min(diff_moderated_deprecated$adj_pval, na.rm = TRUE), digits = 9), 5.1129e-05)
+    expect_equal(round(min(diff_proDA_deprecated$adj_pval, na.rm = TRUE), digits = 5), 0.00125)
   })
 }
 
 if (Sys.getenv("TEST_PROTTI") == "true") {
-  test_that("plot_pval_distribution works", {
-    p <- plot_pval_distribution(
+  test_that("depreciated plot_pval_distribution works", {
+    rlang::with_options(lifecycle_verbosity = "warning", {
+      expect_warning(p <- plot_pval_distribution(
+        diff,
+        peptide,
+        pval
+      ))
+    })
+    expect_is(p, "ggplot")
+    expect_error(print(p), NA)
+  })
+}
+
+if (Sys.getenv("TEST_PROTTI") == "true") {
+  test_that("pval_distribution_plot works", {
+    p <- pval_distribution_plot(
       diff,
       peptide,
       pval
@@ -214,9 +348,72 @@ if (Sys.getenv("TEST_PROTTI") == "true") {
   })
 }
 
-test_that("volcano_protti works", {
+if (Sys.getenv("TEST_PROTTI") == "true") {
+  test_that("deprecated volcano_protti works", {
+    sig_prots <- paste0("protein_", 1:25)
+    rlang::with_options(lifecycle_verbosity = "warning", {
+      expect_warning(p <- volcano_protti(
+        data = diff,
+        grouping = peptide,
+        log2FC = diff,
+        significance = adj_pval,
+        method = "significant",
+        target_column = protein,
+        title = "Test tile",
+        x_axis_label = "test x-Axis",
+        y_axis_label = "test y-Axis",
+        log2FC_cutoff = 1,
+        significance_cutoff = 0.05,
+        interactive = FALSE
+      ))
+    })
+    expect_is(p, "ggplot")
+    expect_error(print(p), NA)
+
+    rlang::with_options(lifecycle_verbosity = "warning", {
+      expect_warning(p_interactive <- volcano_protti(
+        data = diff,
+        grouping = peptide,
+        log2FC = diff,
+        significance = adj_pval,
+        method = "significant",
+        target_column = protein,
+        title = "Test tile",
+        x_axis_label = "test x-Axis",
+        y_axis_label = "test y-Axis",
+        log2FC_cutoff = 1,
+        significance_cutoff = 0.05,
+        interactive = TRUE
+      ))
+    })
+    expect_is(p_interactive, "plotly")
+    expect_error(print(p_interactive), NA)
+
+    rlang::with_options(lifecycle_verbosity = "warning", {
+      expect_warning(p_target <- volcano_protti(
+        data = diff,
+        grouping = peptide,
+        log2FC = diff,
+        significance = adj_pval,
+        method = "target",
+        target = "protein_3",
+        target_column = protein,
+        title = "Test tile",
+        x_axis_label = "test x-Axis",
+        y_axis_label = "test y-Axis",
+        log2FC_cutoff = 1,
+        significance_cutoff = 0.05,
+        interactive = FALSE
+      ))
+    })
+    expect_is(p_target, "ggplot")
+    expect_error(print(p_target), NA)
+  })
+}
+
+test_that("volcano_plot works", {
   sig_prots <- paste0("protein_", 1:25)
-  p <- volcano_protti(
+  p <- volcano_plot(
     data = diff,
     grouping = peptide,
     log2FC = diff,
@@ -234,7 +431,7 @@ test_that("volcano_protti works", {
   expect_error(print(p), NA)
 
   if (Sys.getenv("TEST_PROTTI") == "true") {
-    p_interactive <- volcano_protti(
+    p_interactive <- volcano_plot(
       data = diff,
       grouping = peptide,
       log2FC = diff,
@@ -252,7 +449,7 @@ test_that("volcano_protti works", {
     expect_error(print(p_interactive), NA)
   }
 
-  p_target <- volcano_protti(
+  p_target <- volcano_plot(
     data = diff,
     grouping = peptide,
     log2FC = diff,
@@ -282,8 +479,41 @@ test_that("fit_drc_4p works", {
   expect_equal(round(min(drc_fit$anova_pval, na.rm = TRUE), digits = 6), 0.007297)
 })
 
-test_that("plot_drc_4p works", {
-  p <- plot_drc_4p(
+if (Sys.getenv("TEST_PROTTI") == "true") {
+  test_that("deprecated plot_drc_4p works", {
+    rlang::with_options(lifecycle_verbosity = "warning", {
+      expect_warning(p <- plot_drc_4p(
+        data = drc_fit,
+        grouping = peptide,
+        response = normalised_intensity_log2,
+        dose = concentration,
+        targets = c("peptide_1_2"),
+        unit = "uM",
+        y_axis_name = "test y-Axis"
+      ))
+    })
+
+    expect_is(p, "ggplot")
+    expect_warning(print(p), NA)
+
+    rlang::with_options(lifecycle_verbosity = "warning", {
+      expect_warning(p_facet <- plot_drc_4p(
+        data = drc_fit,
+        grouping = peptide,
+        response = normalised_intensity_log2,
+        dose = concentration,
+        targets = c("peptide_1_2", "peptide_1_1"),
+        unit = "uM"
+      ))
+    })
+
+    expect_is(p_facet, "list")
+    expect_warning(expect_error(print(p_facet), NA))
+  })
+}
+
+test_that("drc_4p_plot works", {
+  p <- drc_4p_plot(
     data = drc_fit,
     grouping = peptide,
     response = normalised_intensity_log2,
@@ -296,7 +526,7 @@ test_that("plot_drc_4p works", {
   expect_is(p, "ggplot")
   expect_warning(print(p), NA)
 
-  p_facet <- plot_drc_4p(
+  p_facet <- drc_4p_plot(
     data = drc_fit,
     grouping = peptide,
     response = normalised_intensity_log2,